Fixes #22 - error in duplicate processing causes by duplicate keys in tags_update
If ignore duplicates was set and a duplicate was removed that had just been
inserted in that scan, creating tags_update records would fail as there were
already records with those keys. In this case the original tags_update records
must be deleted and no new ones created - basically the original tags are
removed from the database as if they never existed.

Note - duplicates are flagged if title, artist, album and track are the same.
The check on title, artist, and album is now case insensitive.

Note - track numbers must be exactly the same, so 01 <> 1.