First stage of adding support for works and virtuals specified in a file.
gettags will now read .sp files (extension can be changed in scan.ini) and
create workvirtuals entries in the database (these are similar to tags entries
(updates are tracked in workvirtual_updates).

An example .sp file format is provided in the virtuals folder.

Changes to workvirtuals are tracked as for tags. Changing a track referred to
in a workvirtual will cause a rescan of the workvirtual too.

Note - workvirtuals are not yet propagated to the database by movetags, that
is stage 2.

Note 2 - logging has been updated in gettags. Logs are written to the logs
folder (exceptions are written to the errors folder).