Fixes #46 - proxy now honours discnumber tag when sorting albums comprised of multiple discs.
If a multi-disc album is merged into a single album name, the proxy will now
serve the tracks in discnumber and then tracknumber order (if the discnumber
tag is populated).

The discnumber tag is already captured in gettags and promoted through in
movetags.

Note that 2 new indexes have been added, to take advantage of those you will
need to run a regen (-r) scan or add the indexes manually:

  create index inxTrackAlbumDiscTrackTitles on tracks (album, discnumber, tracknumber, title)
  create index inxTrackDiscTrackTitles on tracks (discnumber, tracknumber, title)