Fixes #62 - alsa output device for streaming is now identified via a script.
On starting the proxy, it checks for the existence of an alsa.device file.
If it doesn't exist it runs getasla.sh to create it. The file contains the
alsa device to be captured from.

This fix also moves some test files from the main folder into a testcode
folder.