Stage 3 of virtuals/works.
You can now specify in the pycpoint.ini file which indexes virtuals and works
are displayed in. Those settings are honoured at higher levels too - so for
instance if you turn off virtuals in the artists index, artists who only have
virtuals will not be displayed (and genres that only have artists who have
virtuals will not be displayed in the genre index either).

Note that 'use_sorts' has been replaced by 'alternative_index_sorting' - for
now only n will work (full simple and advanced support will be added next).

I've also updated the names of the virtual and work formatting fields in
scan.ini and added more documentation.

I've also fixed a bug in virtuals that caused the tags for a previous virtual
to be cascaded into the next virtual if they were not specified in that
virtual.