Latest phase of virtuals/works:
1) titlesort/albumsort/artistsort/albumartistsort/composersort tags are now
recognised in music files.
2) albumsort/artistsort/albumartistsort/composersort can be overridden in .sp
files.
3) cover art can now be specified in .sp files - if prefer_albumart is set
the coverart will be used at album level in preference to folder or track art.
4) cover art in .sp files can be either image files (using a relative or
absolute path) or point to a music file (which must be in the db).
5) Track order in .sp files is now displayed in preference to track order
from the original album.
6) Various bug fixes relating to virtual/work track selection.

Note that you will have to run databaseupdates/update2.sh on your database
to adjust the schema. You will then need to run a rescan (-r) to repopulate
the tracks part of the database (tags will be retained).