Update for wxPython 3.0
Big refactor (big for me at least!) of wxPython calls to migrate from
2.8 to 3.0.  Caused all sorts of random havok.  Hitting asserts I don't
know how to debug in the Scan and Extract process.  Quickest to
reproduce on a repair call on the scan tab.  Hopefully Mark can take a
look. :) As the assert happens in __core.py, which is well out of my
depth.

Updated pubsub for 3.0

Also added a Minimize to Tray feature.  Double click the systray icon to
bring it back to life.

Massive round of comments so, in two more years, I'll know what the hell
I was doing. :)

Added ignoreini entry to the launch tab so that we can future proof the
combo box drop downs from any ini file that mark may add to the folder.