Added "...processing tags!" in logwindow
To make the GUI not appear so catatonic.  Mark, fwiw, the assert I was
having was because the processing tag output comes from STDOUT so fast,
I don't think the threadmanager can keep up.  This is likely MY issue
with how I wrote the thread manager... a more experienced Python person
could probably come up with something safer, so I am not just ignoring
the tag messages and it seems to have resolved the issue.